Revert "t6006: drop "always" color config tests"
This reverts commit f6bf9e204040645a8c4026738721f82f9dc2f500.

That commit was done primarily to prepare for the weakening
of "always" in e8eb23866a (color: make "always" the same as
"auto" in config, 2017-10-03). But since we've now reverted
e8eb23866a, there's no need for us to remove "-c
color.ui=always" from the tests. And in fact it's a good
idea to restore these tests, to make sure that "always"
continues to work.

Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>