Revert "Merge branch 'en/rename-directory-detection'"
This reverts commit 56d9f5870a376a32d3156aa73cc932b6838b4cd5, reversing
changes made to 5e8fdf623c0a4755be1ddc881e83d17a26fd1d59.

The topic appears to inflict severe regression in renaming merges,
even though the promise of it was that it would improve them.

We do not yet know which exact change in the topic was wrong, but in
the meantime, let's play it safe and revert it out of 'master'
before real Git-using projects are harmed.

Signed-off-by: Junio C Hamano <gitster@pobox.com>